+++ core/src/com/ray3k/skincomposer/dialog/DialogFactory.java
@@ -88,6 +88,15 @@ public class DialogFactory {
         return dialog;
     }
     
+    public DialogExport showDialogExport(DialogListener listener) {
+        var dialog = new DialogExport(main);
+        if (listener != null) {
+            dialog.addListener(listener);
+        }
+        dialog.show(main.getStage());
+        return dialog;
+    }
+    
     public DialogColors showDialogColors(StyleProperty styleProperty,
         DialogColors.DialogColorsListener listener, DialogListener dialogListener) {
         DialogColors dialog = new DialogColors(main, styleProperty, listener);